Create cef-custom.xml
adding custom keys that cef-custom requires to parse the log output messages from rsa malware service in 11.1 as CEF format back to log decoder.

for file events you get 3 hashes, one goes into checksum, the other two are not parsed.  this added them into checksum as well.
User is matched in the cef.xml but that matching is case sensitive and USER from malware logs is not extracted.  this parses that value as well into the same key so that the submitting user gets pulled from the cef logs.

extra hashes are important so that NetWitness Orchestrator can leverage those hashes to look over theat data and determine if they are known.  same with matching against endpoint logs so that common hashes can be located.