Create v20_winevent_nicmsg-custom.cxml
custom file for sysmon updated events to extract the rulename for any v8 rules that have RuleName: entries.  Also normalizes a bunch of rules that were missing the <space> metakey that was breaking a few event.  I have sample logs that can be provided to test the parsing when you are ready to work on this.